* Use diamond type <>. * Remove unnecessary boxing. * Remove unnecessary unboxing. * Replace 'String.indexOf()' expression with 'contains()'. * Replace 'for' loop with enhanced 'for' loop.

+++ jspwiki-main/src/main/java/org/apache/wiki/tags/InsertDiffTag.java
@@ -97,7 +97,7 @@ public class InsertDiffTag extends WikiTagBase {
 
         if( ctx.getPage() != null ) {
             final JspWriter out = pageContext.getOut();
-            final String diff = engine.getManager( DifferenceManager.class ).getDiff( ctx, vernew.intValue(), verold.intValue() );
+            final String diff = engine.getManager( DifferenceManager.class ).getDiff( ctx, vernew, verold);
 
             if( diff.isEmpty() ) {
                 return EVAL_BODY_INCLUDE;
